Add initial zap utilities
The doc.go is the most useful introduction to this work. In addition,
it is built to be extensible, for example it is easy to GetConfig() and
modify it for personal preferences before passing it to BuildForKlog.

This also contains the helper function to allow glog and klog to
coexist.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>