Remove caller skip from BuildForCtrl (#5)
The previous behavior was good for cases where the logger was stored in
a global variable and used directly, _without_ calling `WithName` or
`WithValues` inside a function. However, if one of those methods is
called, the resulting logger will not record the correct function, file,
and line number.

It does not seem feasible to make it work correctly in both cases.

Since using `WithValues` in a function is very useful, and it is not a
big burden to use `WithName` in a function instead of just using a
global logger directly, the new behavior should be better.

The documentation/example have been updated to reflect the proper
usage, and warn about this situation.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>